app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.126.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.38.0",
]
